Rename packages from autoconfig -> autoconfigure
Fixes gh-213
