+++ graphql-dgs-extended-scalars/src/main/kotlin/com/netflix/graphql/dgs/autoconfigure/DgsExtendedScalarsAutoConfiguration.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.netflix.graphql.dgs.autoconfig
+package com.netflix.graphql.dgs.autoconfigure
 
 import com.netflix.graphql.dgs.DgsComponent
 import com.netflix.graphql.dgs.DgsRuntimeWiring
