+++ graphql-dgs-spring-boot-oss-autoconfigure/src/main/resources/META-INF/spring.factories
@@ -1,5 +1,5 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-  com.netflix.graphql.dgs.autoconfig.DgsAutoConfiguration
+  com.netflix.graphql.dgs.autoconfigure.DgsAutoConfiguration
 
 org.springframework.boot.diagnostics.FailureAnalyzer=\
   com.netflix.graphql.dgs.diagnostics.SchemaFailureAnalyzer
