+++ graphql-dgs-spring-boot-oss-autoconfigure/src/test/kotlin/com/netflix/graphql/dgs/autoconfigure/DefaultExceptionHandlerTest.kt
@@ -14,10 +14,10 @@
  * limitations under the License.
  */
 
-package com.netflix.graphql.dgs.autoconfig
+package com.netflix.graphql.dgs.autoconfigure
 
 import com.netflix.graphql.dgs.DgsQueryExecutor
-import com.netflix.graphql.dgs.autoconfig.testcomponents.TestExceptionDatFetcherConfig
+import com.netflix.graphql.dgs.autoconfigure.testcomponents.TestExceptionDatFetcherConfig
 import com.netflix.graphql.dgs.exceptions.QueryException
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
