+++ graphql-dgs-spring-boot-oss-autoconfigure/src/test/kotlin/com/netflix/graphql/dgs/autoconfigure/testcomponents/MyCustomDgsContextBuilder.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.netflix.graphql.dgs.autoconfig.testcomponents
+package com.netflix.graphql.dgs.autoconfigure.testcomponents
 
 import com.netflix.graphql.dgs.DgsComponent
 import com.netflix.graphql.dgs.DgsData
